+++ web/views/component_view.ex
@@ -1,23 +1,5 @@
 defmodule Healthlocker.ComponentView do
   use Healthlocker.Web, :view
-  import Phoenix.HTML.Link
-
-  def button_base, do: "f5 link dim dib mb4 ph3 pv2 br-pill hl-dark-blue"
-  def button_primary, do: button_base() <> " " <> "hl-bg-yellow"
-  def button_secondary, do: button_base() <> " " <> "ba bw1 b--hl-yellow bg-white"
-  def button_inactive, do: button_base() <> " " <> "hl-bg-grey"
-
-  def link_primary(text, opts) do
-    link(text, [class: button_primary()] ++ opts)
-  end
-
-  def link_secondary(text, opts) do
-    link(text, [class: button_secondary()] ++ opts)
-  end
-
-  def link_inactive(text, opts) do
-    link(text, [class: button_inactive()] ++ opts)
-  end
 
   def get_options(option_type) do
     # load security questions from file
